Corrected Copyright year.
The general rule is that the year to include in a copyright notice is the year of first publication of the work. First publication is when the work is made available to the public without restriction.  This includes selling a book or leaving free copies of it in a public place or posting content on your blog. Publication does not include, for example, circulating copies internally to co-workers. For compilations or derivative works which incorporate previously published content, use the year of first publication of the compilation or derivative work.